requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
av==12.3.0
azure-cognitiveservices-speech==1.41.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
cryptography==43.0.1
dacite==1.8.1
dill==0.3.9
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.110.3
fastapi-cli==0.0.5
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.9.0
future==1.0.0
google-ai-generativelanguage==0.6.6
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-crc32c==1.6.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
Jinja2==3.1.4
jiter==0.6.1
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.1
numpy==1.26.4
onnxruntime==1.19.2
openai==1.51.2
outspeed==0.1.147
packaging==24.1
pillow==10.4.0
propcache==0.2.0
proto-plus==1.24.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyee==12.0.0
Pygments==2.18.0
pylibsrtp==0.10.0
pyloudnorm==0.1.1
pyOpenSSL==24.2.1
pyparsing==3.1.4
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
realtime-aioice==0.9.0
requests==2.32.3
rich==13.9.2
rsa==4.9
scipy==1.14.1
sentry-sdk==2.16.0
setuptools==69.5.1
shellingham==1.5.4
silero-vad==5.1.2
sniffio==1.3.1
starlette==0.37.2
sympy==1.13.3
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.29.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==12.0
yarl==1.15.0
Collapse
pandas